requirements.txt
asgiref==3.4.1
Django==3.2.8
djangorestframework==3.12.4
generics==3.4.0
Pillow==8.4.0
pytz==2021.3
sqlparse==0.4.2

django-modeltranslation~=0.17.3